kaggle-polars-downloader/pyproject.toml
[project]
name = "kaggle-polars-downloader"
version = "0.1.0"
description = "Download Kaggle datasets and inspect them with Polars."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kaggle>=1.6.17",
    "polars>=1.28.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
kaggle-polars = "kaggle_polars_downloader.main:main"
csv-to-parquet = "kaggle_polars_downloader.csv_to_parquet:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaggle_polars_downloader"]

[tool.uv]
package = true
